+++ Wrappers/Python/ccpi/viewer/CILViewer.py
@@ -270,8 +270,7 @@ class CILInteractorStyle(vtk.vtkInteractorStyleTrackballCamera):
         planew = vtk.vtkImplicitPlaneWidget2()
 
         rep = vtk.vtkImplicitPlaneRepresentation()
-        world_extent = self.GetImageWorldExtent()
-        extent = [0, world_extent[0], 0, world_extent[1], 0, world_extent[2]]
+        extent = self.GetDataExtentInWorld()
         rep.SetWidgetBounds(*extent)
         planew.SetRepresentation(rep)
         planew.SetInteractor(viewer.getInteractor())
@@ -416,10 +415,39 @@ class CILInteractorStyle(vtk.vtkInteractorStyleTrackballCamera):
         return [(image_coordinates[i]) * spac[i] + orig[i] for i in range(3)]
 
     def GetImageWorldExtent(self):
-        """
-        Compute and return the maximum extent of the image in the rendered world
-        """
+        """Deprecated. Use `GetDataExtentInWorld` and `GetMinMaxVoxelsFromExtent`."""
         return self.image2world(self.GetInputData().GetExtent()[1::2])
+
+    def GetDataExtentInWorld(self):
+        """
+        Compute and return the extent of the input data in world coordinates.
+        """
+        data_extent_image = self.GetInputData().GetExtent()
+        data_extent_world = self.Image2WorldExtent(data_extent_image)
+        return data_extent_world
+
+    def GetMinMaxVoxelsFromExtent(self, extent):
+        """Given the extent of a box or image, gets the voxels corresponding to the min values in all directions
+        and max values in all directions."""
+        voxel_min = extent[0::2]
+        voxel_max = extent[1::2]
+        return voxel_min, voxel_max
+
+    def Image2WorldExtent(self, extent_image):
+        """Given the extent of a box or image, gets the voxels corresponding to the min values in all directions
+        and max values in all directions. Then, converts their coordinates in the world coordinate system. 
+        Returns the converted extent."""
+        voxel_min_image, voxel_max_image = self.GetMinMaxVoxelsFromExtent(extent_image)
+        voxel_min_world = self.image2world(voxel_min_image)
+        voxel_max_world = self.image2world(voxel_max_image)
+        extent_world = self.GetExtentFromVoxels(voxel_min_world, voxel_max_world)
+        return extent_world
+
+    def GetExtentFromVoxels(self, voxel_min, voxel_max):
+        """Given the voxels corresponding to the min values in all directions
+        and max values in all directions, calculates the extent of the box or image they enclose."""
+        extent = (voxel_min[0], voxel_max[0], voxel_min[1], voxel_max[1], voxel_min[2], voxel_max[2])
+        return extent
 
     def GetInputData(self):
         return self._viewer.img3D